requirements.txt
Flask
requests
pycryptodome
urllib3